data_collector/requirements.txt
requests
python-binance
psycopg2-binary